+++ src/Core/Core/Execution/Utilities/ResolverContext.CoerceArgumentValue.cs
@@ -1,15 +1,11 @@
 using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Globalization;
-using System.Threading;
 using System.Threading.Tasks;
 using HotChocolate.Execution.Instrumentation;
 using HotChocolate.Language;
 using HotChocolate.Properties;
 using HotChocolate.Resolvers;
-using HotChocolate.Types;
-using HotChocolate.Utilities;
 
 namespace HotChocolate.Execution
 {
@@ -40,19 +36,26 @@ namespace HotChocolate.Execution
 
             if (typeof(IValueNode).IsAssignableFrom(typeof(T)))
             {
-                if (argumentValue.Literal == null)
-                {
-                    return (T)argumentValue.Type.ParseValue(value);
-                }
-                else
-                {
-                    return (T)argumentValue.Literal;
-                }
+                IValueNode literal =  (argumentValue.Literal == null)
+                    ? argumentValue.Type.ParseValue(value)
+                    : argumentValue.Literal;
+
+                return (T)VariableToValueRewriter.Rewrite(
+                    literal,
+                    Field.Type,
+                    _executionContext.Variables,
+                    _executionContext.Converter);
             }
 
             if (argumentValue.Literal != null)
             {
-                value = argumentValue.Type.ParseLiteral(argumentValue.Literal);
+                IValueNode literal = VariableToValueRewriter.Rewrite(
+                    argumentValue.Literal,
+                    Field.Type,
+                    _executionContext.Variables,
+                    _executionContext.Converter);
+
+                value = argumentValue.Type.ParseLiteral(literal);
             }
 
             if (value is null)
@@ -102,129 +105,5 @@ namespace HotChocolate.Execution
             converted = default;
             return false;
         }
-    }
-
-    public sealed class VariableToValueRewriter
-        : SyntaxRewriter<object>
-    {
-        private static readonly ThreadLocal<VariableToValueRewriter> _current =
-            new ThreadLocal<VariableToValueRewriter>(
-                () => new VariableToValueRewriter());
-        private readonly Stack<IType> _type = new Stack<IType>();
-        public IVariableCollection _variables;
-        public ITypeConversion _typeConversion;
-
-        public IValueNode RewriteValue(
-            IValueNode value,
-            IType type,
-            IVariableCollection variables,
-            ITypeConversion typeConversion)
-        {
-            if (value is null)
-            {
-                throw new ArgumentNullException(nameof(value));
-            }
-
-            if (type is null)
-            {
-                throw new ArgumentNullException(nameof(type));
-            }
-
-            if (variables is null)
-            {
-                throw new ArgumentNullException(nameof(variables));
-            }
-
-            if (typeConversion is null)
-            {
-                throw new ArgumentNullException(nameof(typeConversion));
-            }
-
-            _variables = variables;
-            _typeConversion = typeConversion;
-
-            _type.Clear();
-            _type.Push(type);
-
-            return RewriteValue(value, null);
-        }
-
-        protected override ObjectFieldNode RewriteObjectField(
-            ObjectFieldNode node,
-            object context)
-        {
-            if (_type.Peek().NamedType() is InputObjectType inputObject
-                && inputObject.Fields.TryGetField(
-                    node.Name.Value,
-                    out InputField field))
-            {
-                IValueNode rewritten = null;
-                _type.Push(field.Type);
-
-                switch (node.Value)
-                {
-                    case ListValueNode value:
-                        rewritten = RewriteListValue(value, context);
-                        break;
-
-                    case ObjectValueNode value:
-                        rewritten = RewriteObjectValue(value, context);
-                        break;
-
-                    case VariableNode value:
-
-
-                    default:
-                        rewritten = node.Value;
-                        break;
-                }
-
-                _type.Pop();
-
-                if (rewritten == node.Value)
-                {
-                    return node;
-                }
-                return node.WithValue(rewritten);
-            }
-
-            // TODO : Resources
-            throw new InvalidOperationException("Unknown field type.");
-        }
-
-        protected override ListValueNode RewriteListValue(
-            ListValueNode node,
-            object context)
-        {
-            _type.Push(_type.Peek().ListType().ElementType);
-
-            ListValueNode current = node;
-
-            current = RewriteMany(current, current.Items, context,
-                RewriteValue, current.WithItems);
-
-            _type.Pop();
-
-            return current;
-        }
-
-        private IValueNode ReplaceVariable(VariableValue variableValue)
-        {
-            if (_variables.TryGetVariable(
-                value.Name.Value,
-                out object v))
-            {
-                if (!field.Type.ClrType.IsInstanceOfType(v))
-                {
-
-                }
-            }
-        }
-
-        public static IValueNode Rewrite(
-            IValueNode value,
-            IType type,
-            IVariableCollection variables) =>
-            _current.Value.RewriteValue(value, type, variables);
     }
 }
